+++ data-migrations/2024-10-03-0733-create-affiliations.sql
@@ -1,5 +1,6 @@
 CREATE TABLE `affiliations` (
-  `id` VARCHAR(255) PRIMARY KEY,
+  `id` INT AUTO_INCREMENT PRIMARY KEY,
+  `uri` VARCHAR(255) NOT NULL,
   `provenance` VARCHAR(255) NOT NULL DEFAULT 'DMPTOOL',
   `name` VARCHAR(255) NOT NULL,
   `displayName` VARCHAR(255) NOT NULL,
@@ -19,17 +20,20 @@ CREATE TABLE `affiliations` (
   `feedbackMessage` TEXT,
   `feedbackEmails` JSON DEFAULT NULL,
   `managed` BOOLEAN NOT NULL DEFAULT 0,
+  `active` BOOLEAN NOT NULL DEFAULT 1,
   `createdById` INT NOT NULL,
   `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `modifiedById` INT NOT NULL,
   `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  CONSTRAINT unique_displayName UNIQUE (`displayName`),
+  CONSTRAINT unique_affiliation_uri UNIQUE (`uri`),
+  CONSTRAINT unique_affiliation_displayName UNIQUE (`displayName`),
   FOREIGN KEY (createdById) REFERENCES users(id),
   FOREIGN KEY (modifiedById) REFERENCES users(id),
+  INDEX affiliations_uri_idx (`uri`),
   INDEX affiliations_search_idx (`searchName`),
   INDEX affiliations_sso_idx (`ssoEntityId`),
   INDEX affiliations_funders_idx (`funder`),
-  INDEX affiliations_provenance_idx (`provenance`, `id`, `displayName`)
+  INDEX affiliations_provenance_idx (`provenance`, `uri`, `displayName`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb3;
 
 # Add a unique constraint to the following fields, this approach allows the column to contain NULL
@@ -37,9 +41,9 @@ ALTER TABLE `affiliations`
   ADD UNIQUE KEY unique_affiliation_fundrefId (`fundrefId`);
 
 # The email domains associated with an affiliation to assist with SSO
-CREATE TABLE `affiliations_email_domains` (
+CREATE TABLE `affiliationEmailDomains` (
   `id` INT AUTO_INCREMENT PRIMARY KEY,
-  `affiliationId` VARCHAR(255) NOT NULL,
+  `affiliationId` INT NOT NULL,
   `emailDomain` VARCHAR(255) NOT NULL,
   `createdById` INT NOT NULL,
   `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -51,9 +55,9 @@ CREATE TABLE `affiliations_email_domains` (
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb3;
 
 # The URLs defined by the affiliation admins that are displayed on the sites sub header
-CREATE TABLE `affiliations_links` (
+CREATE TABLE `affiliationLinks` (
   `id` INT AUTO_INCREMENT PRIMARY KEY,
-  `affiliationId` VARCHAR(255) NOT NULL,
+  `affiliationId` INT NOT NULL,
   `url` VARCHAR(255) NOT NULL,
   `text` VARCHAR(255) NOT NULL,
   `createdById` INT NOT NULL,