+++ src/controllers/__tests__/signupController.spec.ts
@@ -21,6 +21,7 @@ const mockedUser: UserModel.User = {
   acceptedTerms: true,
   password: casual.uuid,
   created: new Date().toISOString(),
+  tableIdAsString: false,
   errors: [],
 
   isValid: jest.fn(),